ejemplos/bot-python/requirements.txt
python-telegram-bot==12.0.0
pycrypto==2.6.1
